# gateway/docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.loginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  handlers.refreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  handlers.tokenResponse:
    properties:
      access_token:
        type: string
      expires_in:
        type: integer
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  models.CreateEventRequest:
    properties:
      decision:
        $ref: '#/definitions/models.Decision'
      event_type:
        $ref: '#/definitions/models.EventType'
      matched_signature_id:
        type: string
      payload: {}
      reason:
        type: string
      run_id:
        type: string
      severity:
        $ref: '#/definitions/models.Severity'
      source:
        type: string
    required:
    - decision
    - event_type
    - severity
    - source
    type: object
  models.Decision:
    enum:
    - ALLOWED
    - BLOCKED
    - FLAGGED
    - REDACTED
    type: string
    x-enum-varnames:
    - DecisionAllowed
    - DecisionBlocked
    - DecisionFlagged
    - DecisionRedacted
  models.EventType:
    enum:
    - PROMPT_SCAN
    - TOOL_INTERCEPT
    - OUTPUT_SCAN
    - NETWORK_INTERCEPT
    - FILESYSTEM_INTERCEPT
    - BEHAVIORAL_ALERT
    - HALLUCINATION_DETECTION
    - BROWSER_INTERCEPT
    - CODE_SCAN
    - POLICY_CHANGE
    - RED_TEAM_RUN
    type: string
    x-enum-varnames:
    - EventTypePromptScan
    - EventTypeToolIntercept
    - EventTypeOutputScan
    - EventTypeNetworkIntercept
    - EventTypeFilesystemIntercept
    - EventTypeBehavioralAlert
    - EventTypeHallucinationDetection
    - EventTypeBrowserIntercept
    - EventTypeCodeScan
    - EventTypePolicyChange
    - EventTypeRedTeamRun
  models.SecurityEvent:
    properties:
      decision:
        $ref: '#/definitions/models.Decision'
      event_type:
        $ref: '#/definitions/models.EventType'
      id:
        type: string
      matched_signature_id:
        type: string
      payload: {}
      reason:
        type: string
      run_id:
        type: string
      severity:
        $ref: '#/definitions/models.Severity'
      source:
        type: string
      timestamp:
        type: string
    type: object
  models.Severity:
    enum:
    - INFO
    - LOW
    - MEDIUM
    - HIGH
    - CRITICAL
    type: string
    x-enum-varnames:
    - SeverityInfo
    - SeverityLow
    - SeverityMedium
    - SeverityHigh
    - SeverityCritical
...
host: localhost:8080
info:
  contact: {}
  description: Runtime security platform for autonomous AI agents.
  title: AgentShield API
  version: "1.0"
paths:
  /agents:
    get:
      description: Return all registered agent types from the runtime service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List available agents
      tags:
      - agents
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate with email and password, receive JWT tokens.
      parameters:
      - description: Login credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.tokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Login
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Exchange a valid refresh token for a new access/refresh token pair.
      parameters:
      - description: Refresh token
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.refreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.tokenResponse'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Refresh tokens
      tags:
      - auth
  /events:
    get:
      description: Return up to 500 security events, optionally filtered.
      parameters:
      - description: Filter by run ID
        in: query
        name: run_id
        type: string
      - description: Filter by severity
        in: query
        name: severity
        type: string
      - description: Filter by event type
        in: query
        name: event_type
        type: string
      - description: Filter by decision
        in: query
        name: decision
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.SecurityEvent'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List security events
      tags:
      - events
    post:
      consumes:
      - application/json
      description: Persist a new security event emitted by an interceptor.
      parameters:
      - description: Event payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.CreateEventRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/models.SecurityEvent'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create security event
      tags:
      - events
  /models:
    get:
      description: Return all supported LLM models from the runtime service.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
        "502":
          description: Bad Gateway
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List available models
      tags:
      - agents
...
  /runs/{id}/behavior:
    get:
      description: Return all BEHAVIORAL_ALERT events for the given run.
      parameters:
      - description: Run UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get run behavioral alerts
      tags:
      - runs
  /runs/{id}/hallucination:
    get:
      description: Return the hallucination score and HALLUCINATION_DETECTION events
        for the given run.
      parameters:
      - description: Run UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get run hallucination score
      tags:
      - runs
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"